backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
qdrant-client==1.11.3
python-dotenv==1.0.1
pydantic==2.9.2
 